app/requirements.txt
flask==2.0.3
werkzeug==2.0.3
mysql-connector-python==8.0.26